pyproject.toml
[project]
name = "fmradio"
version = "0.1.0"
description = "FM Radio application using RTL-SDR"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "textual>=0.40.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyrtlsdr>=0.2.93",
    "sounddevice>=0.4.6",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
fmradio = "fmradio.__main__:main"
